moved ResourceType:
