+++ azure_sdk_cosmos/src/requests/get_collection_builder.rs
@@ -1,8 +1,7 @@
-use crate::clients::ResourceType;
 use crate::prelude::*;
 use crate::responses::GetCollectionResponse;
 use crate::CollectionClientRequired;
-use crate::{CollectionClient, CosmosClient, DatabaseClient};
+use crate::{CollectionClient, CosmosClient, DatabaseClient, ResourceType};
 use azure_sdk_core::errors::{check_status_extract_headers_and_body, AzureError};
 use azure_sdk_core::prelude::*;
 use hyper::StatusCode;
